Split Blitz into Blitz-core and Blitz
