+++ blitz-core/src/util.rs
